requirements.txt
beautifulsoup4==4.12.3
joblib==1.3.2
numpy==1.24.3
python_whois==0.9.3
Requests==2.31.0
sitesniffer==0.2
streamlit==1.33.0
tensorflow==2.13.1
tld==0.13
scikit-learn
xgboost
